refactor(sdk): Move tests specs and utils into bento to be shared with e2e tests
